requirements.txt
streamlit==1.27.2
openai==1.3.0
python-dotenv==1.0.0
requests==2.31.0
numpy==1.24.3
scikit-learn==1.3.2
beautifulsoup4==4.12.2
lxml==4.9.3
sentence-transformers==2.2.2
torch>=2.0.0
transformers>=4.30.0